backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
sqlparse==0.5.3
psycopg2==2.9.10
pandas==2.3.2
Pillow==11.3.0
openpyxl==3.1.5
xlrd==2.0.1
PyJWT>=2.9.0
